+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Parallel processing with PSOCK clusters is now more efficient, due to carefully avoiding sending extraneous information to each worker (#384, #396).
 
+* The engine arguments for xgboost `alpha`, `lambda`, and `scale_pos_weight` are now tunable.
+
 # tune 0.1.6
 
 * When using `load_pkgs()`, packages that use random numbers on start-up do not affect the state of the RNG. We also added more control of the RNGkind to make it consistent with the user's previous value (#389).